configurando dockerfile
configurado dockerfile

configurado dockerfile
