+++ TesteAiko.sln
@@ -3,11 +3,16 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.6.30114.105
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api", "src\Api\Api.csproj", "{0E3B167E-87FA-4DB6-A0AF-B29E1235EB6E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Api", "src\Api\Api.csproj", "{0E3B167E-87FA-4DB6-A0AF-B29E1235EB6E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "src\Core\Core.csproj", "{42C41AF9-7F16-4AFE-A24D-6DF50240D8B4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core", "src\Core\Core.csproj", "{42C41AF9-7F16-4AFE-A24D-6DF50240D8B4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Testes", "src\Testes\Testes.csproj", "{B1CD6FA9-1CA4-45C3-AA4C-359455F39206}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "src\Tests\Tests.csproj", "{B1CD6FA9-1CA4-45C3-AA4C-359455F39206}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{7DD4C019-D1B7-4E5A-ABC7-1E61B2EE1416}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,9 +22,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{0E3B167E-87FA-4DB6-A0AF-B29E1235EB6E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -58,5 +60,11 @@ Global
 		{B1CD6FA9-1CA4-45C3-AA4C-359455F39206}.Release|x64.Build.0 = Release|Any CPU
 		{B1CD6FA9-1CA4-45C3-AA4C-359455F39206}.Release|x86.ActiveCfg = Release|Any CPU
 		{B1CD6FA9-1CA4-45C3-AA4C-359455F39206}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {92CE5E67-2B6F-42BA-BC18-07F195F7204F}
 	EndGlobalSection
 EndGlobal
